+++ src/main/java/ec/app/klandscapes/KLandscapes.java
@@ -122,7 +122,7 @@ public class KLandscapes extends GPProblem implements SimpleProblemForm {
             {
             double score = fitness(((GPIndividual) ind).trees[0].child);
             SimpleFitness f = ((SimpleFitness) ind.fitness);
-            f.setFitness(state, score, score==1.0);
+            f.setFitness(state,(float) score, score==1.0);
             ind.evaluated = true;
             }
         
@@ -131,7 +131,7 @@ public class KLandscapes extends GPProblem implements SimpleProblemForm {
     double fitness(GPNode root)
         {
         // Compute the penality (it increases with the difference in depth between the tree and k.
-        double penalty = 1.0/(1+Math.abs(k+1-root.depth()));
+        double penalty = 1/(1+Math.abs(k+1-root.depth()));
         return penalty*fitnessHelper(root)/bestFitness;
         }
